+++ service-api/app/src/App/src/Service/SystemMessage/SystemMessage.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Service;
+namespace App\Service\SystemMessage;
 
 use Aws\Ssm\SsmClient;
