+++ service-api/app/test/AppTest/Service/Aws/SSMClientFactoryTest.php
@@ -2,6 +2,11 @@
 
 declare(strict_types=1);
 
+namespace AppTest\Service\Aws;
+
+use App\Service\Aws\SSMClientFactory;
+use Aws\Sdk;
+use Aws\Ssm\SsmClient;
 use PHPUnit\Framework\ExpectationFailedException;
 use PHPUnit\Framework\MockObject\IncompatibleReturnValueException;
 use PHPUnit\Framework\MockObject\MethodCannotBeConfiguredException;
@@ -11,9 +16,6 @@ use PHPUnit\Framework\MockObject\MethodParametersAlreadyConfiguredException;
 use PHPUnit\Framework\TestCase;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
-use Aws\Sdk;
-use Aws\Ssm\SsmClient;
-use App\Service\Aws\SSMClientFactory;
 use Psr\Container\NotFoundExceptionInterface;
 
 class SSMClientFactoryTest extends TestCase
@@ -26,7 +28,7 @@ class SSMClientFactoryTest extends TestCase
         parent::setUp();
 
         $this->container = $this->createMock(ContainerInterface::class);
-        $this->factory   = new SSMClientFactory();
+        $this->factory = new SSMClientFactory();
     }
 
     /**
@@ -43,7 +45,7 @@ class SSMClientFactoryTest extends TestCase
      */
     public function testInvokeReturnsSsmClient(): void
     {
-        $sdk       = $this->createMock(Sdk::class);
+        $sdk = $this->createMock(Sdk::class);
         $ssmClient = $this->createMock(SsmClient::class);
 
         $sdk->expects($this->once())
